maint: move load_save.py under data_loader because we will reuse it for loading SC (not only fMRI related stuff). To not break imports, just recreate the symlink in docs/analysis/ pointing to the new location
